requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
python-multipart>=0.0.7
websockets>=12.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0 
Pillow>=10.0.0